requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
cinemagoer==2023.5.1
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
IMDbPY==2022.7.9
lxml==5.3.0
motor==3.6.0
pillow==11.0.0
pydub==0.25.1
pymongo==4.9.2
pyshorteners==1.0.1
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
telegraph==2.2.0
typing_extensions==4.12.2
urllib3==2.2.3
